render 함수 구현 세팅
