+++ package.json
@@ -13,6 +13,7 @@
     "@babel/core": "^7.26.0",
     "@babel/preset-env": "^7.26.0",
     "@babel/preset-react": "^7.26.3",
+    "@types/node": "^22.10.2",
     "typescript": "~5.6.2",
     "vite": "^6.0.3"
   },
