+++ src/components/Counter.tsx
@@ -1,0 +1,10 @@
+const Counter = ({ count }: { count: number }) => {
+  return (
+    <div className='count-container'>
+      <p>{count}</p>
+      <p>{count}</p>
+    </div>
+  );
+};
+
+export default Counter;
