+++ src/main.tsx
@@ -1,46 +1,10 @@
-const App = () => {
-  let count = 0;
+import App from './App';
+import render from './my_react/render';
 
-  const handleIncrease = () => {
-    count += 1;
-  };
+const $app = document.getElementById('app');
 
-  const handleDecrease = () => {
-    count -= 1;
-  };
-
-  return (
-    <>
-      <div></div>
-      <div>
-        <h1>Counter</h1>
-        <Counter count={count} />
-        <Button onIncrease={handleIncrease} onDecrease={handleDecrease} />
-      </div>
-    </>
-  );
-};
-
-const Counter = ({ count }: { count: number }) => {
-  return (
-    <div className='count-container'>
-      <p>{count}</p>
-      <p>{count}</p>
-    </div>
-  );
-};
-
-const Button = ({ onIncrease, onDecrease }: { onIncrease: () => void; onDecrease: () => void }) => {
-  return (
-    <div className='button-container'>
-      <button className='increase' onClick={onIncrease}>
-        +
-      </button>
-      <button className='decrease' onClick={onDecrease}>
-        -
-      </button>
-    </div>
-  );
-};
+if ($app instanceof HTMLElement) {
+  render($app, App());
+}
 
 console.log(JSON.stringify(App(), null, 2));
