+++ src/my_react/render.ts
@@ -1,0 +1,5 @@
+const render = (root: HTMLElement, vDom: JSX.Element) => {
+  console.log(root, vDom);
+};
+
+export default render;
